test: separate rateLimit tests into unit and integration tests
